Flags fixes + add channelOverride flag (#14)
Adding a channelOverride flag. This is useful for those who wants to run
this application side by side their existing solution. This way all
messages can be overwritten to a single channel, for debugging purposes
(i.e. do I get the same messages in my monolith channel as my original
channels?).

It also can be used if one would like to _force_ all users to only allow
sending to a specific channel. This way one could setup this proxy with
their unique endpoint. I.e.

my-channel.this-app with `-channelOverride #my-channel` and make it a
limited proxy.